Store dependency info in a separate file, not in the chunk morphology
This means that the tool can work without requiring changes to Morph to
allow new fields in morphologies. It also makes it easier to throw away
the foreign dependency information once an import is complete.

The downside is that there are now 3 types of extension, not two, and
a lot of code needed factoring out into base classes (and more needs
factoring out still). There are now .to_lorry, .to_chunk and .find_deps
extensions.