omnibus: Recommend `--deployment` flag for `bundle install`
This avoids a situation where `bundle install` expects a bunch of stuff
to exist in /usr or /home but you have since moved the repo around.
Installing stuff in /usr is not recommended in Baserock because it gets
lost on upgrades (and /usr will one day be read only, I expect).