Fix rubygems.to_chunk failing when run inside `bundle exec`
By calling Bundler::Dsl.gemspec without specifying :path, this tool
was causing the Bundler::Source::Path instance for the target .gemspec
to be for path '.', which is relative path that is only valid inside
one Dir.chdir block. It seems that all the Bundler resolution code runs
inside that block and so there should be no problem, but unless we
specify an absolute path for the gemspec then errors like this sometimes
appear:

    /usr/lib/ruby/gems/2.0.0/gems/bundler-1.7.6/lib/bundler/resolver.rb:357:in
    `resolve': Could not find gem 'ffi-yajl (>= 0) ruby' in source at ..
    (Bundler::GemNotFound)
    Source does not contain any versions of 'ffi-yajl (>= 0) ruby'

This normally happens when an Omnibus import chains to rubygems.to_chunk
for a RubyGem component.

The path is clearly valid at the time Bundler::Dsl.gemspec is called
(otherwise it'd raise a Bundler::InvalidOption exception at the time)
but not valid later (hence the error). Note that the second '.' in that
error message is a full stop and not part of the path!