Give better errors when one .lorry duplicates another one
Now the tool gives the filenames of both files involved in the conflict.

The most likely reason for seeing this error is that, at least in
the lorries.git on git.baserock.org[1], the toplevel directory contains
some 'disabled' subdirectories, so the user needs to point --lorries-dir
to the correct subdirectory so that duplicate disabled lorries aren't
loaded.

This isn't really the final word on that problem, because having the
tool load disabled lorries is quite bad and if there are no conflicts
the user won't notice that it's happening and other confusing stuff
might occur. But hopefully it's good enough for now.

1. http://git.baserock.org/cgi-bin/cgit.cgi/baserock/local-config/lorries.git/tree/