rubygems: Detect more signed Gems
The multi_json Gem wasn't being detected as signed, because the
'signing_key' field is an expression that can evaluate to 'nil' in some
cases. In this Gem the 'cert_chain' field was still a standard string.
Hopefully checking for the presence of either will catch all cases
(and false positives should be harmless anyway).